Added Carthage and MaterialKit side view controller
