+++ BodyweightFitness-Bridging-Header.h
@@ -1,0 +1,13 @@
+//
+//  BodyweightFitness-Bridging-Header.h
+//  BodyweightFitness
+//
+//  Created by Damian Mazurkiewicz on 05/01/2016.
+//  Copyright © 2016 Damian Mazurkiewicz. All rights reserved.
+//
+
+#ifndef BodyweightFitness_Bridging_Header_h
+#define BodyweightFitness_Bridging_Header_h
+
+
+#endif /* BodyweightFitness_Bridging_Header_h */
