+++ BodyweightFitness/Frameworks/AnimatedFrame.swift
@@ -1,4 +1,6 @@
 /// Keeps a reference to an `UIImage` instance and its duration as a GIF frame.
+import UIKit
+
 struct AnimatedFrame {
     let image: UIImage?
     let duration: NSTimeInterval
