+++ Cartfile.resolved
@@ -1,0 +1,2 @@
+github "CosmicMind/MaterialKit" "1.27.4"
+github "SwiftyJSON/SwiftyJSON" "2.3.3"
