--- Crashlytics.framework/Info.plist
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>BuildMachineOSBuild</key>
-	<string>14E46</string>
-	<key>CFBundleDevelopmentRegion</key>
-	<string>English</string>
-	<key>CFBundleExecutable</key>
-	<string>Crashlytics</string>
-	<key>CFBundleIdentifier</key>
-	<string>com.twitter.crashlytics.ios</string>
-	<key>CFBundleInfoDictionaryVersion</key>
-	<string>6.0</string>
-	<key>CFBundleName</key>
-	<string>Crashlytics</string>
-	<key>CFBundlePackageType</key>
-	<string>FMWK</string>
-	<key>CFBundleShortVersionString</key>
-	<string>3.3.1</string>
-	<key>CFBundleSignature</key>
-	<string>????</string>
-	<key>CFBundleSupportedPlatforms</key>
-	<array>
-		<string>iPhoneOS</string>
-	</array>
-	<key>CFBundleVersion</key>
-	<string>79</string>
-	<key>DTCompiler</key>
-	<string>com.apple.compilers.llvm.clang.1_0</string>
-	<key>DTPlatformBuild</key>
-	<string>13A340</string>
-	<key>DTPlatformName</key>
-	<string>iphoneos</string>
-	<key>DTPlatformVersion</key>
-	<string>9.0</string>
-	<key>DTSDKBuild</key>
-	<string>13A340</string>
-	<key>DTSDKName</key>
-	<string>iphoneos9.0</string>
-	<key>DTXcode</key>
-	<string>0700</string>
-	<key>DTXcodeBuild</key>
-	<string>7A218</string>
-	<key>MinimumOSVersion</key>
-	<string>6.0</string>
-	<key>NSHumanReadableCopyright</key>
-	<string>Copyright © 2015 Crashlytics, Inc. All rights reserved.</string>
-	<key>UIDeviceFamily</key>
-	<array>
-		<integer>1</integer>
-		<integer>2</integer>
-	</array>
-</dict>
-</plist>
