--- Crashlytics.framework/Modules/module.modulemap
@@ -1,13 +1,0 @@
-framework module Crashlytics {
-    header "Crashlytics.h"
-    header "Answers.h"
-    header "CLSLogging.h"
-    header "CLSReport.h"
-    header "CLSStackFrame.h"
-    header "CLSAttributes.h"
-
-    export *
-
-    link "z"
-    link "c++"
-}
