--- Fabric.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module Fabric {
-    umbrella header "Fabric.h"
-
-    export *
-    module * { export * }
-}
