+++ MainViewController.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+class MainViewController: UINavigationController {
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
